Use central Structure class
The central Structure class acts as a proxy for
optimade.models.StructureResource, as well as a lazy converter into
other structure classes and types.
At the moment a Structure can be converted into an AiiDA StructureData,
an ASE Atoms, and a pymatgen Structure/Molecule (depending on the
dimension_types, Structure for bulk 3D structures and Molecule for all
others).

The central Structure class is now used throughout the widget instead of
ASE Atoms and/or AiiDA StructureData.