Start making a template for GitHub issues
Specifically for issues made when using the link in the application to
report bugs/issues. At this point in time, this is only publicly
available from MyBinder, hence the specific naming of it. This should be
changed in the future. The actual issue body should be changed as well
to include the debug-level log file.