Functional report buttons
Since the ipywidgets do not support other HTML Button elements than of
type "button", is was necessary to "manually" create the report buttons
in an HTML widget.
The suggestion report button is simply a form + "submit" type button.
The bug report button is a bit more sophisticated, with a but of
JavaScript directly in the value of the "onclick" attribute.
It retrieves a truncated version of the debug logs and passes them
directly into an issue via GitHub's query parameters.