Add more downloadable file formats
Open the possibility to go via ASE.Atoms or to convert an OPTIMADE
structure to another format. Add some well-known formats.

Extend with custom warnings classes, as well as a way of "wrapping"
other warnings in order to ensure logging.

Add ASE to install requirements. (As well as Voilà).

Add functionality to handle pymatgen.[Structure/Molecule] as well.
But since the conversion is very specific to either class type, it has
not been opened up to use pymatgen in the current download list.

Furthermore, there doesn't seem to be anything gained by having both
ASE and pymatgen.