Introduce range sliders as filter inputs
Specifically add IntRangeSliders for `nelements` and `nsites`.
All of the infrastructure for adding other than Text widgets for filter
inputs has been added.

Other things:
Try to make `msg` the first argument for all Exceptions as best as
possible. Currently the only exception to this is BadResource, where it
is needed to pass a resource.