Use periodic table widget for elements input
Wrap PTableWidget in a custom widget.
This makes it easier to use in FilterInputs as well as to toggle HAS ANY
or HAS ALL filterings.

Wrap `collect_value` in a try-except in order to properly raise if an
exception occurs.

Allow the filter parser to return operators.
OPERATORS_MAP has become a fallback, with the default operator being
`""`.