Use CacheControl and requests.Session.
First, make sure request URLs are similar to previous request URLs.
All query parameters and values are either added in an OrderedDict or
extracted from a complete URL, sorted and reencoded into a complete URL.

Using FileCache:
Store cached requests responses in an object store-like cache using
cachecontrol[filecache], i.e., the lockfile package.
Expiration is set to 1 day.

Add buttons to clear cache and logs.
These are only visible if in debug mode (can be toggled on or off in the
log accordion).

Setup separate CacheControlAdapters.
In this way, caching can be avoided for requests to `http://localhost`
or `http://127.0.0.1`.