Reset intrangeslider and results pager
Reset the results pager "silently"; it seems
`self.hold_trait_notification()` doesn't work for the normal `reset()`
method, hence a new method `silent_reset()` has been created, which
utilizes `set_pagination_data()` to "reset" the page chooser.
In this way, the widget is not suddenly updated, which results in the
first chosen database to not function properly.

IntRangerSlider's were not reset for databases with no content, this is
now the case, using a "defaults" mapper.

A status text with "No structures found!" is also added, to give the
user some feedback concerning the result, since it otherwise seems like
nothing happened.