Add button_style parameter to all format() calls
The new button_style variable was not supplied in several calls to the
`format()` function when setting the download button's HTML.

This also adds an automatic default for the `ButtonStyle` Enum
("DEFAULT").
A warning will be logged (and shown in the log under normal conditions)
that the supplied button_style value was wrong, and it has been set to
"DEFAULT", if it was wrong.

button_style type of ButtonStyle is also allowed now.