Revisit dependencies, add "server" extra
Remove the following (unused) dependencies:
- appmode
- jupyterlab
- numpy

Moved the following dependencies to the new `server` extra:
- ase
- voila

Any non-native OPTIMADE adapter formats have been demoted to only be
included if the package is installed. This means the download format
widget can be used without having ASE installed.
For a more complete experience, ASE is kept as a dependency for the
`server` extra.

Make the CLI work if Voilà is not installed.

Update README with new dependency setup.
Create `Contribute` section in README.