Use ipywidgets-extended DropdownExtended widget
Disable providers that cannot be used with the client, but keep them in
the Dropdown of providers.
The disabled providers cannot be chosen due utilizing the
DropdownExtended widget, which let's you individually disable options.
Sort provider list to attach "disabled" providers at the end of the
options list.
Skip "exmpl", "optimade", and "aiida" permanently from the OPTIMADE list
of providers.