Add parameter for skipping/disabling providers+dbs
This makes it possible to skip specific providers (based on their `id`).

It also makes it possible to skip specific databases (based on their
`id`) for any provider.

The value must be a dict, where the keys are the provider name (NB! Not
id) and the values are a list of child database ids.