Use `grouping` for child_db dropdown throughout
New parameter for `ProviderImplementationChooser`, i.e., for
`OptimadeQueryProviderWidget`: `provider_database_groupings`.
It must be a dictionary where the keys are the database labels/names
(NOT ids!).
The values are then another dictionary, where the keys are the group
headings and values are lists of database ids.

If any database id is not listed, it is added at the end outside a
category.