Use ase format for viewer and default download format
The structure adaptor of optimate-python-tooll now not properly deal
with the 2d structures in pdb and cif format which always purge cell.
This cause the issue that viewer will not shouw cell for two-dimensional
structures and the download format cif broken.
The ase format as adapter works fine, and it is also the adaptor in
aiidalab-base-widget and works all good there.